third_party/blink/web_tests/platform/generic/fast/forms/elements-invalidate-on-form-attribute-invalidation-expected.txt

Tests form.elements is invalidated when input element's form attribute is changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

collection = document.getElementById('someForm').elements;
PASS collection.length is 1
PASS collection.length; input.setAttribute('form', 'otherForm'); collection.length is 0
PASS collection.length; input.setAttribute('form', 'someForm'); collection.length is 1
PASS successfullyParsed is true

TEST COMPLETE
